+++ src/ClassLibrary1/CL1AppModule.cs
@@ -1,0 +1,9 @@
+﻿using Reface.AppStarter.AppModules;
+
+namespace ClassLibrary1
+{
+    [ComponentScanAppModule]
+    public class CL1AppModule : AppModule
+    {
+    }
+}