+++ src/ClassLibrary1/Services/ICL1Service.cs
@@ -1,0 +1,7 @@
+﻿namespace ClassLibrary1.Services
+{
+    public interface ICL1Service
+    {
+        string GetName();
+    }
+}