+++ src/Reface.App.Starter/AppModules/ComponentScanAppModule.cs
@@ -1,6 +1,5 @@
 ﻿using Reface.AppStarter.AppContainerBuilders;
 using Reface.AppStarter.Attributes;
-using Reface.AppStarter.AutofacExt;
 using System;
 using System.Linq;
 using System.Reflection;
@@ -12,12 +11,28 @@ namespace Reface.AppStarter.AppModules
     /// </summary>
     public class ComponentScanAppModule : AppModule
     {
+        public class MethodAndAttrInfo<TAttr> where TAttr : Attribute
+        {
+            public MethodInfo Method { get; private set; }
+            public TAttr Attribute { get; private set; }
+
+            public MethodAndAttrInfo(MethodInfo method, TAttr attribute)
+            {
+                Method = method;
+                Attribute = attribute;
+            }
+        }
+
         public override void OnUsing(AppSetup setup, IAppModule targetModule)
         {
             AutofacContainerBuilder autofacContainerBuilder
                 = setup.GetAppContainerBuilder<AutofacContainerBuilder>();
             RegisterScanResult(setup, targetModule, autofacContainerBuilder);
             RegisterComponentFromMethods(targetModule, autofacContainerBuilder);
+            setup.AllModulesLoaded += (sender, e) =>
+              {
+                  ReplaceComponentFromMethods(targetModule, autofacContainerBuilder);
+              };
         }
 
         /// <summary>
@@ -30,19 +45,45 @@ namespace Reface.AppStarter.AppModules
         {
             var methods = targetModule.GetType().GetMethods()
                 .Where(x => x.ReturnType != typeof(void))
-                .Select(x => new
-                {
-                    Method = x,
-                    Attr = x.GetCustomAttribute<ComponentCreatorAttribute>()
-                })
-                .Where(x => x.Attr != null);
+                .Select(x => new MethodAndAttrInfo<ComponentCreatorAttribute>(x, x.GetCustomAttribute<ComponentCreatorAttribute>()))
+                .Where(x => x.Attribute != null)
+                .Select(x => x.Method);
+            RegisterMethods(targetModule, autofacContainerBuilder, methods);
+        }        /// <summary>
+                 /// 从 <see cref="IAppModule"/> 实例的方法中创建组件，
+                 /// 这些方法必须被标记 <see cref="ComponentAttribute"/> 特征。
+                 /// </summary>
+                 /// <param name="targetModule"></param>
+                 /// <param name="autofacContainerBuilder"></param>
+        private static void ReplaceComponentFromMethods(IAppModule targetModule, AutofacContainerBuilder autofacContainerBuilder)
+        {
+            var methods = targetModule.GetType().GetMethods()
+                .Where(x => x.ReturnType != typeof(void))
+                .Select(x => new MethodAndAttrInfo<ReplaceCreatorAttribute>(x, x.GetCustomAttribute<ReplaceCreatorAttribute>()))
+                .Where(x => x.Attribute != null)
+                .Select(x => x.Method);
+            foreach (var method in methods)
+            {
+                autofacContainerBuilder.RemoveComponentByServiceType(method.ReturnType);
+            }
+            RegisterMethods(targetModule, autofacContainerBuilder, methods);
+        }
+
+        /// <summary>
+        /// 注册所有方法
+        /// </summary>
+        /// <param name="targetModule"></param>
+        /// <param name="autofacContainerBuilder"></param>
+        /// <param name="methods"></param>
+        private static void RegisterMethods(IAppModule targetModule, AutofacContainerBuilder autofacContainerBuilder, System.Collections.Generic.IEnumerable<MethodInfo> methods)
+        {
             foreach (var method in methods)
             {
                 autofacContainerBuilder.RegisterByCreator(cm =>
                 {
-                    ParameterInfo[] ps = method.Method.GetParameters();
+                    ParameterInfo[] ps = method.GetParameters();
                     if (ps.Length == 0)
-                        return method.Method.Invoke(targetModule, null);
+                        return method.Invoke(targetModule, null);
                     object[] values = new object[ps.Length];
                     for (int i = 0; i < ps.Length; i++)
                     {
@@ -50,8 +91,8 @@ namespace Reface.AppStarter.AppModules
                         object value = cm.CreateComponent(pType);
                         values[i] = value;
                     }
-                    return method.Method.Invoke(targetModule, values);
-                }, method.Method.ReturnType);
+                    return method.Invoke(targetModule, values);
+                }, method.ReturnType);
             }
         }