+++ src/Reface.App.Starter/Attributes/ComponentAttribute.cs
@@ -9,6 +9,7 @@ namespace Reface.AppStarter.Attributes
     public class ComponentAttribute
         : ScannableAttribute
     {
+
         public RegistionMode RegistionMode { get; private set; }
 
         public ComponentAttribute(RegistionMode registionMode = RegistionMode.AsInterfaces)