+++ src/Reface.AppStarter.Tests/TestAppModule.cs
@@ -1,4 +1,7 @@
-﻿using Reface.AppStarter.AppModules;
+﻿using ClassLibrary1;
+using ClassLibrary1.Services;
+using Reface.AppStarter.AppContainerBuilders;
+using Reface.AppStarter.AppModules;
 using Reface.AppStarter.Attributes;
 using Reface.AppStarter.Tests.AppContainerBuilders;
 using Reface.AppStarter.Tests.Services;
@@ -7,6 +10,7 @@ namespace Reface.AppStarter.Tests
 {
     [AutoConfigAppModule]
     [ComponentScanAppModule]
+    [CL1AppModule]
     public class TestAppModule : AppModule
     {
         public override void OnUsing(AppSetup setup, IAppModule targetModule)
@@ -19,6 +23,12 @@ namespace Reface.AppStarter.Tests
         public IServiceRegistedByAppModule GetServiceRegistedByAppModule()
         {
             return new DefaultServiceRegistedByAppModule();
+        }
+
+        [ReplaceCreator]
+        public ICL1Service GetCL1Service()
+        {
+            return new MyDefaultCL1Service();
         }
     }
 }